consensus/parlia: use unit64 to express epochLength and blockInterval

+++ consensus/parlia/ramanujanfork.go
@@ -29,7 +29,7 @@ func (p *Parlia) delayForRamanujanFork(snap *Snapshot, header *types.Header) tim
 }
 
 func (p *Parlia) blockTimeForRamanujanFork(snap *Snapshot, header, parent *types.Header) uint64 {
-	blockTime := parent.MilliTimestamp() + uint64(snap.BlockInterval)
+	blockTime := parent.MilliTimestamp() + snap.BlockInterval
 	if p.chainConfig.IsRamanujan(header.Number) {
 		blockTime = blockTime + p.backOffTime(snap, parent, header, p.val)
 	}
@@ -42,7 +42,7 @@ func (p *Parlia) blockTimeForRamanujanFork(snap *Snapshot, header, parent *types
 
 func (p *Parlia) blockTimeVerifyForRamanujanFork(snap *Snapshot, header, parent *types.Header) error {
 	if p.chainConfig.IsRamanujan(header.Number) {
-		if header.MilliTimestamp() < parent.MilliTimestamp()+uint64(snap.BlockInterval)+p.backOffTime(snap, parent, header, header.Coinbase) {
+		if header.MilliTimestamp() < parent.MilliTimestamp()+snap.BlockInterval+p.backOffTime(snap, parent, header, header.Coinbase) {
 			return consensus.ErrFutureBlock
 		}
 	}
